+++ ali/common-lootjs/src/test/java/com/yanny/ali/lootjs/test/LootJsConditionTooltipTest.java
@@ -177,7 +177,7 @@ public class LootJsConditionTooltipTest {
     public void testMatchAnyInventorySlotTooltip() {
         assertTooltip(LootJsConditionTooltipUtils.matchAnyInventorySlot(UTILS, new MatchAnyInventorySlot(ItemFilter.ENCHANTED, true)).build(), List.of(
                 "Match Any Inventory Slot:",
-                "  -> Predicate:",
+                "  -> Item Filter: ENCHANTED",
                 "  -> Hotbar: true"
         ));
     }